Cast wave number to float in enemy hp scaling

+++ world/World.cpp
@@ -989,7 +989,7 @@ void World::addEntity(std::shared_ptr<Entity> entity, bool defer) {
 
     if (entity->isEnemy()) {
         float playerDamageMultiplier = _player->getDamageMultiplier();
-        entity->setMaxHitPoints(entity->getMaxHitPoints() + (entity->getMaxHitPoints() * (playerDamageMultiplier / 2)) + ((getCurrentWaveNumber() / 4)));
+        entity->setMaxHitPoints(entity->getMaxHitPoints() + (entity->getMaxHitPoints() * (playerDamageMultiplier / 2)) + (((float)getCurrentWaveNumber() / 4)));
         entity->heal(entity->getMaxHitPoints());
 
         _enemies.push_back(entity);
